Remove SyncClient and add sync methods to AdbClient
